Added a helpfull error message:
previously errors due to not having a locator assigned to a app-map node and errors due to protractor being unable to construct an "element" object were undistinguised.
With the change, if a user puts a malformed xpath in, and protractors validation catches it, that errer will be passed out instead of hidden.